test: 기능 구현으로 인해 깨지는 레거시 테스트 교정

+++ backend/src/main/java/edonymyeon/backend/post/domain/Post.java
@@ -68,7 +68,7 @@ public class Post {
     private LocalDateTime createdAt;
 
     @ColumnDefault("0")
-    private Long viewCount = 0L;
+    private Integer viewCount = 0;
 
     public Post(
             final String title,
